Sell Out - Reel Big Fish (#110)
- Add license headers
  - Use licensure to automatically apply them and check in ci
- Apply some clippy lints
- Switch to nextest for better test integration with Jenkins
  - Can output compatible JUnit XML test results

+++ nxrm_two_portal/src/config.rs
@@ -1,3 +1,6 @@
+// Copyright (c) 2024-present Sonatype, Inc. All rights reserved.
+// "Sonatype" is a trademark of Sonatype, Inc.
+
 use config::{Config, Environment};
 use portal_api::CENTRAL_HOST;
 use serde::Deserialize;
@@ -13,7 +16,7 @@ impl AppConfig {
         let env_source = Environment::with_prefix("nxrm_two_portal");
         let app_config = Config::builder()
             .set_default("central_url", CENTRAL_HOST)?
-            .set_default("app_port", 2727 as u16)?
+            .set_default("app_port", 2727_u16)?
             .add_source(env_source)
             .build()?
             .try_deserialize()?;
